--- .package.json
@@ -1,88 +1,0 @@
-{
-    "scripts": {
-        "start": "ts-node ./server/main.ts watch stdin",
-        "manager": "ts-node -P tsconfig.json ./scripts/manager.ts",
-        "build": "ts-node ./server/bundler.ts",
-        "test": "ts-node ./scripts/tests/all-tests.ts",
-        "init": "ts-node ./scripts/init.ts"
-    },
-    "devDependencies": {
-        "@popperjs/core": "^2.11.8",
-        "@tsconfig/deno": "^1.0.9",
-        "@tsconfig/svelte": "^5.0.2",
-        "@types/bad-words": "^3.0.3",
-        "@types/busboy": "^1.5.3",
-        "@types/callsite": "^1.0.34",
-        "@types/datatables.net-fixedheader": "^3.1.8",
-        "@types/express-session": "^1.18.0",
-        "@types/fuzzy-search": "^2.1.5",
-        "@types/inquirer": "^9.0.7",
-        "@types/jquery": "^3.5.29",
-        "@types/keypress": "^2.0.30",
-        "@types/multer": "^1.4.11",
-        "@types/node": "^20.11.24",
-        "@types/nodemailer": "^6.4.14",
-        "@types/objects-to-csv": "^1.3.3",
-        "@types/pg": "^8.11.2",
-        "@types/prompt": "^1.1.8",
-        "@types/prompts": "^2.4.9",
-        "@types/uuid": "^9.0.8",
-        "@typescript-eslint/eslint-plugin": "^6.19.0",
-        "@typescript-eslint/parser": "^6.19.0",
-        "callsite": "^1.0.0",
-        "esbuild": "0.20.1",
-        "eslint": "^8.56.0",
-        "prettier": "^3.2.2",
-        "prettier-plugin-svelte": "^3.1.2",
-        "svelte": "^4.2.8",
-        "svelte-check": "^3.6.3",
-        "svelte-preprocess-esbuild": "^3.0.1",
-        "svelte-recaptcha-v2": "^0.0.2",
-        "tslib": "^2.6.2",
-        "typescript": "^5.3.3"
-    },
-    "dependencies": {
-        "@beyowi/svelte-recaptcha-v2": "^0.0.4",
-        "@fortawesome/fontawesome-free": "^6.5.1",
-        "@neoxia-js/nodemailer-sendgrid-transport": "^0.1.1",
-        "@putout/cli-choose": "^2.0.0",
-        "@types/bootstrap": "^5.2.10",
-        "@types/jqueryui": "^1.12.21",
-        "@types/nodemailer-sendgrid": "^1.0.3",
-        "axios": "^1.6.7",
-        "bad-words": "^3.0.4",
-        "bootstrap": "^5.3.3",
-        "bootstrap-icons": "^1.11.3",
-        "busboy": "^1.6.0",
-        "cli-select": "^1.1.2",
-        "cliffy": "^2.5.1",
-        "compress-images": "^1.9.9",
-        "deep-email-validator": "^0.1.21",
-        "directory-tree": "^3.5.1",
-        "dotenv": "^16.4.5",
-        "esbuild-svelte": "^0.8.0",
-        "express": "^4.18.3",
-        "express-session": "^1.18.0",
-        "fuzzy-search": "^3.2.1",
-        "icons": "^1.0.0",
-        "inquirer": "^9.2.15",
-        "jest": "^29.7.0",
-        "jquery": "^3.7.1",
-        "keypress": "^0.2.1",
-        "material-icons": "^1.13.12",
-        "multer": "^1.4.5-lts.1",
-        "node-html-constructor": "^4.0.3",
-        "nodemailer": "^6.9.11",
-        "nodemailer-sendgrid-transport": "^0.2.0",
-        "objects-to-csv": "^1.3.6",
-        "pg": "^8.11.3",
-        "prompt": "^1.3.0",
-        "prompts": "^2.4.2",
-        "socket.io": "^4.7.4",
-        "socket.io-client": "^4.7.4",
-        "test": "^3.3.0",
-        "ts-node": "^10.9.2",
-        "uuid": "^9.0.1",
-        "webcam-easy": "^1.1.1"
-    }
-}
