+++ client/entries/test/matchVideo.ts
@@ -1,26 +1,26 @@
 import { FIRSTEvent } from '../../models/FIRST/event';
 
-(async() => {
+(async () => {
     // to pull
     const result = await FIRSTEvent.current?.getMatches();
     if (result && result.isOk()) {
-    const matches = result.value; // FIRSTMatch[]
+        const matches = result.value; // FIRSTMatch[]
         for (const match of matches) {
-        match.tba // here is the tba information in the exact structure listed above
+            match.tba; // here is the tba information in the exact structure listed above
         }
     }
 
-    const video = document.createElement("video");
+    const video = document.createElement('video');
     video.setAttribute('controls', 'true');
     video.setAttribute('width', '250');
 
     const src1 = document.createElement('source');
-    src1.src = "/media/cc0-videos/flower.webm";
+    src1.src = '/media/cc0-videos/flower.webm';
     src1.type = 'video/webm';
 
     video.append(src1);
 
-`
+    `
 <video controls width="250">
   <source src="/media/cc0-videos/flower.webm" type="video/webm" />
 
@@ -32,13 +32,5 @@ import { FIRSTEvent } from '../../models/FIRST/event';
   <a href="/media/cc0-videos/flower.mp4">MP4</a>
   video.
 </video>
-`
-
-
-
-
-
-
-
-
+`;
 })();
