--- client/views/components/robot-display/matchVideos.svelte
@@ -1,29 +1,0 @@
-<script lang="ts">
-
-import {FIRSTMatch} from '../../../models/FIRST/match'
-export let match: FIRSTMatch | undefined;
-let matchVideos: string[] = [];
-
-const pullVideos = (m: FIRSTMatch | undefined) => {
-    if (m) {
-    matchVideos = m.tba.videos
-      .filter(v => v.type === 'youtube') // I'm guessing with this, you'll likely have to log it to see
-        .map(v => v.key);
-    } else {
-    matchVideos = [];
-    }
-}
-
-$: pullVideos(match);
-
-
-
-</script>
-{#each matchVideos as video}
-<iframe
-    width="560" height="315" src="https://www.youtube.com/embed/{video}"
-    title="YouTube video player" frameborder="0" 
-    allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture; web-share" 
-    referrerpolicy="strict-origin-when-cross-origin" allowfullscreen>
-</iframe>
-{/each}
